requirements.txt
#Django
Django==2.1
django-cors-headers==3.1.0
django-redis==4.10.0
djangorestframework==3.10.2
djangorestframework-api-key==2.0.0
djangorestframework-jsonapi==2.8.0
djangorestframework-simplejwt==4.3.0
django-filter==2.2.0
django-debug-toolbar==2.0

# utilities
redis==3.3.11
pymysql

# Testing
mypy==0.782
pytest==6.0.1
pytest-sugar==0.9.4
pytest-django==3.9.0
coverage==6.1.1
factory_boy==3.2.0
